Merged the one and only James Baxter into main

+++ src/main/java/org/acitech/entities/Enemy.java
@@ -49,8 +49,8 @@ public class Enemy extends Entity {
         if (this.position.distance(playerPos) < aggroDistance) {
             this.acceleration = new Vector2D(x, y);
             this.acceleration = this.acceleration.scalarMultiply(moveSpeed);
-            if (this.position.distance(playerPos) < (int) (this.width/2) ||
-                    this.position.distance(playerPos) < (int) (this.height/2)) {
+            if (this.position.distance(playerPos) < ((double) this.width /2) ||
+                    this.position.distance(playerPos) < ((double) this.height /2)) {
                 if (GamePanel.player.damageTimer == 0) {
                     GamePanel.player.health -= Math.max(this.damage - GamePanel.player.meleeDefense, 0);
                     GamePanel.player.damageTimer = immunity;
